+++ examples/resources/ctyun_ebm/resource.tf
@@ -1,21 +1,21 @@
 terraform {
   required_providers {
     ctyun = {
-      source = "www.ctyun.cn/ctyun-it/ctyun"
+      source = "ctyun-it/ctyun"
     }
   }
 }
 
 provider "ctyun" {
-  region_id            = "bb9fdb42056f11eda1610242ac110002"         # 如果此值不填，则默认读取环境变量中的CTYUN_REGION_ID
-  az_name              = "cn-huadong1-jsnj1A-public-ctcloud"        # 如果此值不填，则默认读取环境变量中的CTYUN_AZ_NAME
-  env                  = "prod"                                     # 如果此值不填，则默认读取环境变量中的CTYUN_ENV
+  region_id            = "bb9fdb42056f11eda1610242ac110002"
+  az_name              = "cn-huadong1-jsnj1A-public-ctcloud"
+  env                  = "prod"
 }
 
 resource "ctyun_ebm" "ebm_test" {
   device_type = "physical.s5.2xlarge4"
   instance_name = "ebm-0312-tf"
-  hostname = "ebm-0310-tf"
+  hostname = "ebm-0317-tf"
   image_uuid = "im-xevpi6apqilz1bixmogofyref9qm"
   password = "P@ss132345"
   security_group_id = "sg-vrp4x1lm7p"