+++ CHANGELOG.md
@@ -1,2 +1,23 @@
-# Dev（2024.5.15）
+# Changelog
 
+
+## New Resources
+
+* ctyun_ebm
+* ctyun_ecs
+
+## New Data Sources
+
+* ctyun_ebm_device_types
+
+## Enhancements
+
+* provider: Support Ebm SDK
+
+## Bug Fixes
+
+* resource/ctyun_ecs fix config
+
+## Deprecations
+
+* data-source/ctyun_abc